# evaluacion-ems/src/main/java/com/evaluacion/dto/ResponseDTO.java
package com.evaluacion.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class ResponseDTO {

    private Boolean success;
    private String message;
    private Object responseData;
}